+++ src/ServiceProvider.php
@@ -6,6 +6,7 @@ namespace Ka4ivan\ApiDebugger;
 
 use Illuminate\Http\Request;
 use Illuminate\Routing\Router;
+use Illuminate\Support\Facades\DB;
 use Ka4ivan\ApiDebugger\Middleware\ApiDebuggerMiddleware;
 use Ka4ivan\ApiDebugger\Support\ApiDebugger;
 
@@ -18,11 +19,20 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
      */
     public function boot(): void
     {
+        $this->publishConfig();
         $this->registerMacros();
         $this->registerMiddleware();
 
-        if (app(ApiDebugger::class)->isActive()) {
-            $this->startDebug();
+        if (class_exists(\Laravel\Octane\Events\RequestReceived::class)) {
+            $this->app['events']->listen(\Laravel\Octane\Events\RequestReceived::class, function () {
+                app(ApiDebugger::class)->reset();
+            });
+        }
+
+        if (class_exists(\Laravel\Octane\Events\RequestTerminated::class)) {
+            $this->app['events']->listen(\Laravel\Octane\Events\RequestTerminated::class, function () {
+                app(ApiDebugger::class)->reset();
+            });
         }
     }
 
@@ -33,7 +43,16 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
      */
     public function register(): void
     {
-        $this->app->singleton(ApiDebugger::class);
+        $this->app->scoped(ApiDebugger::class, function () {
+            return new ApiDebugger();
+        });
+    }
+
+    protected function publishConfig(): void
+    {
+        $this->publishes([
+            __DIR__ . '/../config/api-debugger.php' => config_path('api-debugger.php'),
+        ], 'api-debugger-config');
     }
 
     /**